Resolve CredScan issues with Core, KeyVault
Resolves #23505 and resolves #23502
