+++ eng/CredScanSuppression.json
@@ -32,6 +32,7 @@
                 "security",
                 "abcde",
                 "BaR@123e2etests-res-f978ade9",
+                "test_secret",
                 "testPassword1!",
                 "sdkpass",
                 "sdkpass1",
@@ -50,11 +51,13 @@
         },
         {
             "file":[
+                "sdk/core/Azure.Core/tests/PemReaderTests.Keys.cs",
                 "sdk/identity/Azure.Identity/tests/Data/cert.pfx",
                 "sdk/identity/Azure.Identity/tests/Data/cert-invalid-data.pem",
                 "sdk/identity/Azure.Identity/tests/Data/cert-password-protected.pfx",
                 "sdk/identity/Azure.Identity/tests/Data/cert.pem",
                 "sdk/identity/Azure.Identity/tests/Data/mock-arc-mi-key.key",
+                "sdk/keyvault/Azure.Security.KeyVault.Certificates/tests/PemReaderTests.Keys.cs",
                 "sdk/storage/Microsoft.Azure.WebJobs.Extensions.Storage.Common/tests/cert.pem"
             ],
             "_justification": "File contains private key used by test code."
