+++ sdk/core/Azure.Core/tests/Azure.Core.Tests.csproj
@@ -45,4 +45,7 @@
     <Compile Include="..\src\Shared\TypeBinder.cs" LinkBase="Shared" /> 
     <Compile Include="..\..\..\..\common\ManagementTestShared\Redesign\*.cs" LinkBase="Shared\Mgmt" />
   </ItemGroup>
+  <ItemGroup>
+    <Compile Update="PemReaderTests.Keys.cs" DependentUpon="PemReaderTests.cs" />
+  </ItemGroup>
 </Project>
